drop version tag from log4crc config files
The log4crc version tag is optional and is only tested by log4c as a
sanity check (if a version is found in the log4crc file then it must
match the version of log4c).

The presence of a version tag in the RDK log4crc files is blocking
the migration from log4c v1.2.3 to v1.2.4 (the default version of
log4c in OE 2.2).

Change-Id: I5ea8d581e9aa6300a90f7d2be43f488ce3592725
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>