RDKCOM-225: drop version tag from log4crc config files
The log4crc version tag is optional and is only tested by log4c as a
sanity check (if a version is found in the log4crc file then it must
match the version of log4c).
The presence of a version tag in the RDK log4crc files is blocking
the migration from log4c v1.2.3 to v1.2.4 (the default version of
log4c in OE 2.2).
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

Reason for change: drop version tag from log4crc config files
Test Procedure: None.
Risks: Low.

Change-Id: Ia67cced07ad1b1f6ead8689333b91ebdaf2db88f